Hide ecomm_totalvalue If Not Applicable
With the current implementation, the tag assistant shows the error "Number field should not be quoted: 'ecomm_totalvalue'" on pages where the value is empty. The reason is that it is rendered with an empty value `''` and not hidden. This pull request ensures that the value is not rendered at all on pages where it is not applicable.